refactor airtable models
