+++ esimsrouter/esims_router/constants.py
@@ -7,17 +7,11 @@ class RouterConst:
     """RouterConst"""
 
     # Env Variables
-    ENTRIES = "ENTRIES"
     STATE_KEY = "LAMBDA_STATE_KEY"
 
     # Lambda states
     ON = "ON"
     OFF = "OFF"
-
-    # Table Names
-    CARRIERS_TABLE_NAME = "CARRIERS_TABLE_NAME"
-    ATTACHMENT_TABLE_NAME = "ATTACHMENT_TABLE_NAME"
-    FOLDER_FIELD = "Name"
 
     # Root Folder
     DBX_PATH = "/ESims for Gaza/Fresh Sims to Load in Airtable/{}"
