+++ lib/esimslib/connectors/__init__.py
@@ -1,10 +1,4 @@
 """Connectors module"""
 
 from esimslib.connectors.aws_connector import SSMConnector, S3Connector
-from esimslib.connectors.airtable_connector import (
-    AirTableConnector,
-    Providers,
-    Attachments,
-    Donations,
-)
 from esimslib.connectors.dropbox_connector import DropboxConnector
